Fix entity ID suppressor to prevent duplicate IDs and ensure valid entity ID assignment, Updated Upstream Purpur, Fix build

+++ horizon-server/src/main/java/dev/gideonwhite1029/horizon/protocol/syncmatica/SyncmaticaProtocol.java
@@ -103,7 +103,7 @@ public class SyncmaticaProtocol {
         return sanitized.toString().replaceAll(ILLEGAL_PATTERNS, "_");
     }
 
-    public static boolean isOverQuota(int sent) {
+    public static boolean isOverQuota(long sent) {
         return HorizonConfig.syncmaticaQuota && sent > HorizonConfig.syncmaticaQuotaLimit;
     }
